Rejected from app store, added commas to beats and 2x multi to store
The app got rejected for its first submission to the app store. It was for a "lack of content" so I am adding more. I made it so the beats had commas in it on the main page, and I also added a 2x multiplier you can buy to the store that allows for your multiplier to always start at 2