Shop Title Update and prison background
I changed the top of the shop to look better with a VStack of the shop image in red and a smaller points shower. I also added the prison background that allows for 2x and something to strive for further down the games path at 5M points. Added more auto beaters so the 10th and last one is now 10 million beats. Also added black shadow and forced the text on the home screen to be white