Working on Implementing Hit Visual
I found when you get over 100k and it starts going to 100k instead of 100,000 then its not as satisfying for the user. So I am trying to add a animation to where when you hit the meat a + whatever your number is at the time flys off of the meat. It is my first experience with SwifUI animation and I have been stuck